fixes for 64 bit OS X build (c/o david robillard); tested on Lion & Tiger
git-svn-id: svn://localhost/ardour2/branches/3.0@11585 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/engine_dialog.cc
@@ -728,7 +728,7 @@ EngineControl::enumerate_coreaudio_devices ()
 	// (code snippet gently "borrowed" from St?hane Letz jackdmp;)
 	OSStatus err;
 	Boolean isWritable;
-	size_t outSize = sizeof(isWritable);
+	UInt32 outSize = sizeof(isWritable);
 
 	backend_devs.clear ();
 
@@ -744,7 +744,7 @@ EngineControl::enumerate_coreaudio_devices ()
 		if (err == noErr) {
 			// Look for the CoreAudio device name...
 			char coreDeviceName[256];
-			size_t nameSize;
+			UInt32 nameSize;
 
 			for (int i = 0; i < numCoreDevices; i++) {
